Fix issue with switching the theme
- Issue: Theme resets to default theme when navigated to a different page
- Created a helper called getInitialTheme and passed it as initial value to the ThemeProvider
