+++ src/pages/_app.tsx
@@ -1,3 +1,4 @@
+import getInitialTheme from '@/helpers/get-initial-theme'
 import type { AppProps } from 'next/app'
 import ThemeProvider from 'src/context/ThemeContext'
 
@@ -5,7 +6,7 @@ import '../styles/globals.css'
 
 function MyApp({ Component, pageProps }: AppProps) {
   return (
-    <ThemeProvider initialTheme="light">
+    <ThemeProvider initialTheme={getInitialTheme()}>
       <Component {...pageProps} />
     </ThemeProvider>
   )
